pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "voice_assistant"
version = "0.1.0"

description = "A voice assistant with STT, TTS, grammar correction, and LLM capabilities"
readme = "README.md"
requires-python = "3.10"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn>=0.22.0",
    "python-multipart",
    "pydantic>=2.0.0",
    "python-dotenv",
    "transformers>=4.30.0",
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "numpy>=1.24.0",
    "soundfile>=0.12.0",
    "librosa>=0.10.0",
    "language-tool-python>=2.7.1",
    "ctranslate2>=3.17.0",
    "sentencepiece>=0.1.99",
    "cython>=0.29.0"
]

[tool.hatch.metadata]
allow-direct-references = true
